FPLA-852: Removed to follow document status validation as per AC update
