+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validators/interfaces/HasAttachedDocument.java
@@ -1,6 +1,6 @@
 package uk.gov.hmcts.reform.fpl.validators.interfaces;
 
-import uk.gov.hmcts.reform.fpl.validators.HasDocumentStatusValidator;
+import uk.gov.hmcts.reform.fpl.validators.HasAttachedDocumentValidator;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -11,8 +11,8 @@ import javax.validation.Payload;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { HasDocumentStatusValidator.class })
-public @interface HasDocumentStatus {
+@Constraint(validatedBy = { HasAttachedDocumentValidator.class })
+public @interface HasAttachedDocument {
     String message() default "Attach the document or change the status from 'Attached'.";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
